Make the viewer standalone

+++ apps/viewer/cypress/stop.sh
@@ -1,5 +1,12 @@
 #! /bin/sh
-appname=$1
-dockername="cypress_testing_$appname"
+# RUN THIS SCRIPT FROM THE ROOT FOLDER OF YOUR APP
+appname=${PWD##*/}
 
-docker kill $dockername
+if [[ $appname == "cypress" ]]
+then
+	echo "Please run this app from your app root folder."
+else
+	echo "Killing server for the $appname app"
+	dockername="cypress_testing_$appname"
+	docker kill $dockername
+fi
